# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Personal finance management API for tracking transactions, budgets, savings goals, and debts.",
        "title": "WealthPath API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@wealthpath.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/debts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all debts for the current user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debts"
                ],
                "summary": "List debts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_wealthpath_backend_internal_model.Debt"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new debt entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debts"
                ],
                "summary": "Create a debt",
                "parameters": [
                    {
                        "description": "Debt data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_wealthpath_backend_internal_service.CreateDebtInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_wealthpath_backend_internal_model.Debt"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/debts/calculator": {
            "get": {
                "description": "Calculate interest and payment schedule for a loan",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debts"
                ],
                "summary": "Calculate loan interest",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Loan principal amount",
                        "name": "principal",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Annual interest rate (percentage)",
                        "name": "interestRate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Loan term in months",
                        "name": "termMonths",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "fixed",
                        "description": "Payment type (fixed or interest_only)",
                        "name": "paymentType",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_wealthpath_backend_internal_service.InterestCalculatorResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/debts/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a debt by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debts"
                ],
                "summary": "Get a debt",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Debt ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_wealthpath_backend_internal_model.Debt"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing debt",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debts"
                ],
                "summary": "Update a debt",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Debt ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated debt data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_wealthpath_backend_internal_service.UpdateDebtInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_wealthpath_backend_internal_model.Debt"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a debt by ID",
                "tags": [
                    "debts"
                ],
                "summary": "Delete a debt",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Debt ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/debts/{id}/payment": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Record a payment against a debt",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debts"
                ],
                "summary": "Make a debt payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Debt ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payment data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_wealthpath_backend_internal_service.MakePaymentInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_wealthpath_backend_internal_model.Debt"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/debts/{id}/payoff-plan": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Calculate a payoff plan for a debt",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debts"
                ],
                "summary": "Get debt payoff plan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Debt ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Monthly payment amount",
                        "name": "monthlyPayment",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_wealthpath_backend_internal_model.PayoffPlan"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_wealthpath_backend_internal_model.AmortizationRow": {
            "type": "object",
            "properties": {
                "interest": {
                    "type": "number"
                },
                "month": {
                    "type": "integer"
                },
                "payment": {
                    "type": "number"
                },
                "principal": {
                    "type": "number"
                },
                "remainingBalance": {
                    "type": "number"
                }
            }
        },
...
        "github_com_wealthpath_backend_internal_model.Debt": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "currentBalance": {
                    "type": "number"
                },
                "dueDay": {
                    "description": "Day of month",
                    "type": "integer"
                },
                "expectedPayoff": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "interestRate": {
                    "description": "APR as percentage",
                    "type": "number"
                },
                "minimumPayment": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "originalAmount": {
                    "type": "number"
                },
                "startDate": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/github_com_wealthpath_backend_internal_model.DebtType"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "github_com_wealthpath_backend_internal_model.DebtType": {
            "type": "string",
            "enum": [
                "mortgage",
                "auto_loan",
                "student_loan",
                "credit_card",
                "personal_loan",
                "other"
            ],
            "x-enum-varnames": [
                "DebtTypeMortgage",
                "DebtTypeAutoLoan",
                "DebtTypeStudentLoan",
                "DebtTypeCreditCard",
                "DebtTypePersonalLoan",
                "DebtTypeOther"
            ]
        },
...
        "github_com_wealthpath_backend_internal_model.PayoffPlan": {
            "type": "object",
            "properties": {
                "amortizationPlan": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_wealthpath_backend_internal_model.AmortizationRow"
                    }
                },
                "currentBalance": {
                    "type": "number"
                },
                "debtId": {
                    "type": "string"
                },
                "monthlyPayment": {
                    "type": "number"
                },
                "monthsToPayoff": {
                    "type": "integer"
                },
                "payoffDate": {
                    "type": "string"
                },
                "totalInterest": {
                    "type": "number"
                },
                "totalPayment": {
                    "type": "number"
                }
            }
        },
...
        "github_com_wealthpath_backend_internal_service.CreateDebtInput": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string"
                },
                "currentBalance": {
                    "type": "number"
                },
                "dueDay": {
                    "type": "integer"
                },
                "interestRate": {
                    "description": "APR as percentage (e.g., 5.5 for 5.5%)",
                    "type": "number"
                },
                "minimumPayment": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "originalAmount": {
                    "type": "number"
                },
                "startDate": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/github_com_wealthpath_backend_internal_model.DebtType"
                }
            }
        },
...
        "github_com_wealthpath_backend_internal_service.InterestCalculatorResult": {
            "type": "object",
            "properties": {
                "monthlyPayment": {
                    "type": "number"
                },
                "payoffDate": {
                    "type": "string"
                },
                "totalInterest": {
                    "type": "number"
                },
                "totalPayment": {
                    "type": "number"
                }
            }
        },
...
        "github_com_wealthpath_backend_internal_service.MakePaymentInput": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "date": {
                    "type": "string"
                }
            }
        },
...
        "github_com_wealthpath_backend_internal_service.UpdateDebtInput": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string"
                },
                "currentBalance": {
                    "type": "number"
                },
                "dueDay": {
                    "type": "integer"
                },
                "interestRate": {
                    "type": "number"
                },
                "minimumPayment": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "originalAmount": {
                    "type": "number"
                },
                "startDate": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/github_com_wealthpath_backend_internal_model.DebtType"
                }
            }
        },
...
        "internal_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}